refactor: move doulbe origin to kilt-support
